+++ app/src/main/java/com/nextcloud/talk/shareditems/adapters/SharedItemsAdapter.kt
@@ -37,18 +37,19 @@ import com.nextcloud.talk.shareditems.model.SharedItem
 import com.nextcloud.talk.shareditems.model.SharedLocationItem
 import com.nextcloud.talk.shareditems.model.SharedOtherItem
 import com.nextcloud.talk.shareditems.model.SharedPollItem
+import com.nextcloud.talk.ui.theme.ViewThemeUtils
 
 class SharedItemsAdapter(
     private val showGrid: Boolean,
     private val user: User,
     private val roomToken: String,
-    private val isUserConversationOwnerOrModerator: Boolean
+    private val isUserConversationOwnerOrModerator: Boolean,
+    private val viewThemeUtils: ViewThemeUtils
 ) : RecyclerView.Adapter<SharedItemsViewHolder>() {
 
     var items: List<SharedItem> = emptyList()
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): SharedItemsViewHolder {
-
         return if (showGrid) {
             SharedItemsGridViewHolder(
                 SharedItemGridBinding.inflate(
@@ -56,7 +57,8 @@ class SharedItemsAdapter(
                     parent,
                     false
                 ),
-                user
+                user,
+                viewThemeUtils
             )
         } else {
             SharedItemsListViewHolder(
@@ -65,7 +67,8 @@ class SharedItemsAdapter(
                     parent,
                     false
                 ),
-                user
+                user,
+                viewThemeUtils
             )
         }
     }
